+++ schedule_kun_frontend/pages/teachers/lessons/edit/[lessonId].tsx
@@ -1,5 +1,5 @@
 import ContainerButton from '@/components/elements/containerButton';
-import TimePicker from '@/components/elements/timepicker';
+import TimeSpanSelector from '@/components/elements/timeSpanSelector';
 import Dashboard from '@/components/layouts/dashboard';
 import { TeachersClient } from '@/lib/ScheduleKunApi/TeachersClient';
 import type { NextPage } from 'next';
@@ -127,18 +127,13 @@ const EditLesson: NextPage = () => {
                 setLesson({ ...lesson, description: e.target.value });
               }} />
             </FormGroup>
-            <div className={`d-flex ${!!errors?.start_time ? 'is-invalid' : ''}`}>
-              <FormGroup className='d-flex me-3' style={{ width: '200px' }}>
-                <Form.Label className='d-flex align-items-center mb-0' style={{ width: '100px' }}>開始時刻</Form.Label>
-                <TimePicker isInvalid={!!errors?.start_time} style={{ width: '150px' }} time={startTime} onChange={setStartTime} />
-              </FormGroup>
-              <span className='d-flex align-items-center mb-0 me-3'>~</span>
-              <FormGroup className='d-flex me-3' style={{ width: '200px' }}>
-                <Form.Label className='d-flex align-items-center mb-0' style={{ width: '100px' }}>終了時刻</Form.Label>
-                <TimePicker isInvalid={!!errors?.start_time} style={{ width: '150px' }} time={endTime} onChange={setEndTime} />
-              </FormGroup>
-            </div>
-            <Form.Control.Feedback type="invalid">{errors?.start_time}</Form.Control.Feedback>
+            <TimeSpanSelector
+              startTime={startTime}
+              endTime={endTime}
+              setStartTime={setStartTime}
+              setEndTime={setEndTime}
+              errors={errors}
+            />
           </Form>
         </Card>
         <Button className='me-3' onClick={() => setShow(true)}>更新</Button>
